requirements.txt
fastapi[standard]>=0.111.0
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.13.0
noisereduce>=3.0.0
numpy>=1.26.0
matplotlib>=3.8.0
pyyaml>=6.0.0
python-multipart>=0.0.9
aiofiles>=23.0.0
python-dotenv>=1.0.0
